refactor: float pipeline, namespace DSP, Parser factory, grouped structs
- Replace int16 audioData with float [-1,1] in WavParser; decode all
  PCM bit depths and IEEE float formats at parse time
- Move Parser method bodies from header to WavParser.cpp; add static
  Parser::fromFile factory to prevent half-initialised instances
- Convert FFTProcessor class to namespace fft free functions
- Convert WindowFunctions class to namespace window free functions;
  apply() now takes vector<float>&
- GainProcessor::apply updated to float, clamps to [-1,1]
- Add FFTReport and ProcessingOptions structs; flatten removed
- processWavFile takes ProcessingOptions default struct
- writeWavFile converts float to int16 at write boundary only
- Update Commands.cpp, WavGUI.cpp, addon.cpp to new APIs
- Update all three test files: namespace call sites, float assertions,
  remove int16 window overload tests
- Add src/utils/WavParser.cpp to CMakeLists.txt and binding.gyp

+++ include/audio/GainProcessor.hpp
@@ -1,13 +1,12 @@
 #pragma once
 
-#include <cstdint>
 #include <vector>
 
 class GainProcessor {
   public:
     explicit GainProcessor(double gain) : gain_(gain) {}
 
-    void apply(std::vector<int16_t> &samples) const;
+    void apply(std::vector<float> &samples) const;
     [[nodiscard]] double gain() const { return gain_; }
 
   private:
